flask_backend/requirements.txt
Flask==3.0.3
flask-swagger-ui==4.11.1
flask-cors==4.0.1
pymongo
openai==1.41.1
python-dotenv==1.0.1
gunicorn==23.0.0